+++ resources/release-notes.md
@@ -5,6 +5,7 @@
 
 ## Software
 
+- (All) Add `mag.mixin.TimeSeriesOperationSupport` for simplified operations support
 - (All) Fix issues with colors and legends for multiple stackedplots
 - (IMAP) Check there is data before enabling auto-compression event in `mag.imap.view.Field`